requirement.txt
streamlit
langchain
langchain-openai
pypdf
faiss-cpu
python-dotenv
langchain-community
langchain-core